rename: classRegister components 폴더로 컴포넌트 이동 (#206)
