+++ src/pages/instructor/classRegister/components/ClassAmount/index.css.ts
