+++ src/pages/instructor/classRegister/components/ClassGenre/index.css.ts
