+++ src/pages/instructor/classRegister/components/ClassName/index.css.ts
