+++ src/pages/instructor/classRegister/components/ClassPlace/index.css.ts
