+++ src/pages/instructor/classRegister/components/ClassRecommend/index.css.ts
