+++ src/pages/instructor/classRegister/components/ClassRepresentImage/index.css.ts
