+++ src/pages/instructor/classRegister/components/ClassSchedule/TimeStep/TimeControl/index.tsx
