+++ src/pages/instructor/classRegister/components/ClassSchedule/index.css.ts
